fix(mcp-server): simplify runtime mode handoff and diagnostics
